fix: :bug: Ease user existence check as new oauth users were not being able to login
There was an issue while saving new user data to db. here was a check that user is in
session AND in db. Well that is not the case for new users and check failed.
After the fix just check the existence in session and update with that information at hand to db.